#152: created baseSearchServer to allow smart parsing for backends and not redirections
