+++ api/src/main/java/com/rbmhtechnology/vind/api/CompletableSearchServer.java
@@ -3,12 +3,19 @@ package com.rbmhtechnology.vind.api;
 import com.rbmhtechnology.vind.annotations.AnnotationUtil;
 import com.rbmhtechnology.vind.api.query.FulltextSearch;
 import com.rbmhtechnology.vind.api.query.delete.Delete;
-import com.rbmhtechnology.vind.api.query.filter.Filter;
 import com.rbmhtechnology.vind.api.query.get.RealTimeGet;
 import com.rbmhtechnology.vind.api.query.inverseSearch.InverseSearch;
 import com.rbmhtechnology.vind.api.query.suggestion.ExecutableSuggestionSearch;
 import com.rbmhtechnology.vind.api.query.update.Update;
-import com.rbmhtechnology.vind.api.result.*;
+import com.rbmhtechnology.vind.api.result.BeanGetResult;
+import com.rbmhtechnology.vind.api.result.BeanSearchResult;
+import com.rbmhtechnology.vind.api.result.DeleteResult;
+import com.rbmhtechnology.vind.api.result.GetResult;
+import com.rbmhtechnology.vind.api.result.IndexResult;
+import com.rbmhtechnology.vind.api.result.InverseSearchResult;
+import com.rbmhtechnology.vind.api.result.SearchResult;
+import com.rbmhtechnology.vind.api.result.StatusResult;
+import com.rbmhtechnology.vind.api.result.SuggestionResult;
 import com.rbmhtechnology.vind.configure.SearchConfiguration;
 import com.rbmhtechnology.vind.model.DocumentFactory;
 import com.rbmhtechnology.vind.model.InverseSearchQuery;
@@ -290,12 +297,12 @@ public class CompletableSearchServer extends SearchServer {
     }
 
     @Override
-    protected   <T> BeanSearchResult<T> doExecute(FulltextSearch search, Class<T> c) {
+    public <T> BeanSearchResult<T> execute(FulltextSearch search, Class<T> c) {
         return backend.execute(search, c);
     }
 
     @Override
-    protected SearchResult doExecute(FulltextSearch search, DocumentFactory factory) {
+    public SearchResult execute(FulltextSearch search, DocumentFactory factory) {
         return backend.execute(search, factory);
     }
