+++ api/src/main/java/com/rbmhtechnology/vind/api/MasterSlaveSearchServer.java
@@ -118,12 +118,12 @@ public class MasterSlaveSearchServer extends SearchServer {
     }
 
     @Override
-    protected  <T> BeanSearchResult<T> doExecute(FulltextSearch search, Class<T> c) {
+    public  <T> BeanSearchResult<T> execute(FulltextSearch search, Class<T> c) {
         return backend.execute(search, c);
     }
 
     @Override
-    protected SearchResult doExecute(FulltextSearch search, DocumentFactory factory) {
+    public SearchResult execute(FulltextSearch search, DocumentFactory factory) {
         return backend.execute(search, factory);
     }
